+++ src/main/java/lotto/domain/LottoNumberPool.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.data.LottoNumber;
+
 import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
