+++ src/main/java/lotto/domain/data/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.data;
 
 import java.util.Objects;
