+++ src/main/java/lotto/domain/generator/AutoLottoGenerator.java
@@ -1,4 +1,8 @@
-package lotto.domain;
+package lotto.domain.generator;
+
+import lotto.domain.LottoNumberPool;
+import lotto.domain.data.Lotto;
+import lotto.domain.data.LottoNumber;
 
 import java.util.ArrayList;
 import java.util.Collections;
