+++ src/test/java/lotto/domain/LottoNumberPoolTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.data.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
