+++ src/test/java/lotto/domain/LottoStoreTest.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.domain.data.Lotto;
+import lotto.domain.data.ManualLottoMemo;
+import lotto.domain.data.PriceLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
