+++ src/test/java/lotto/domain/data/WinningLottoTest.java
@@ -1,5 +1,9 @@
-package lotto.domain;
+package lotto.domain.data;
 
+import lotto.domain.LottoNumberPool;
+import lotto.domain.data.Lotto;
+import lotto.domain.data.LottoNumber;
+import lotto.domain.data.WinningLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
