+++ src/test/java/lotto/domain/generator/LottoGeneratorTest.java
@@ -1,13 +1,8 @@
-package lotto.domain;
+package lotto.domain.generator;
 
-import lotto.domain.Lotto;
-import lotto.domain.LottoGenerator;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.List;
 
 public class LottoGeneratorTest {
